reduce the default size of the image cache
Reduce the default image cache size to reduce memory
usage an likelihood of a crash due to out of memory.
The new default of 40 is a good balance of cache hits
to memory consumption, results in good reduction of
flicker while zooming in/out while avoiding consumption
of too much memory. Example app now runs with about 260MB