reduce memory usage and improve paint behaviour
Reduce memory usage by reducing image cache size.
Image cache is less useful now that we are scaling
tile widgets as zooming occurs.

Improve painting in mixed mode so that tiles are
painted with vector data first if it is available.
This eliminates raster rendering in many cases,
making the map more efficient because it has
less work to do.

Introduce a new option, tileDelay, so that it
is easier to visualize tile behaviour when loading
tiles.