eliminate tile retention while zooming
Eliminate retention of tiles while zooming
so that the only tiles shown are for the
current zoom. This reintroduces flicker
when new tiles are introduced, which is a
bit jarring for the user. The idea is that
rendering tiles larger than the current
canvas can trigger a memory issue that
causes RSS to grow unreasonably large.