fix bug when multiple tile sources are used
Executor deduplication key is used to avoid
processing the same work in quick succession.
The tile source was not included in the deduplication
key which could cause one vector tile data to
be used for a different source. In practice this
was evident when using hillshade in a separate
tileset, where the hillshade would display for
some tiles without other features. By including
the tile source in the deduplication key, tile
data is appropriately segregated preventing
this issue.