fix race condition causing tile not to appear
When using a background theme, a race condition
would occasionally result in a cancelled tile
request in the background to cancel a tile request
for the foreground, only when zooming. This would
result in the foreground tile not being rendered,
leaving a blank space in the map.

The issue was a map of futures intended to reduce
duplicated work, which remained after executor
cancellation was introduced. A cancelled future
would result in cancelling the wrong request
since they would share the same future due to
the future cache.

Fixed the issue by eliminating shared futures for
cancellable tasks. A future is still shared for
loading tile bytes, however decoding the tile now
relies on executor deduplication key to eliminate
duplicated work.