remove support for mixed and raster mode rendering
Remove support for rendering map tiles
as images.

RenderMode (raster,mixed) was created as
a way to work around performance problems
that caused jank in the UI, especially
while zooming. Both of these modes render
tiles to raster images, which are expensive
to create, occupy a lot of memory, and
require significant disk space for caching.

In recent testing mixed-mode rendering
degrades the user experience significantly
by delaying time-to-first-render.
Mixed and raster render modes are intended
to improve performance but require a lot
more computation. What makes it even worse
is that raster images must be rendered on
the UI thread, leaving less time for
rendering UI frames.
In the case of mixed mode, apps must render
both vectors and images, more than doubling
the work required.

By removing support for raster-based
rendering we will no longer require users
to understand what it is and how to choose
a RenderMode option, or to decide how much
memory and disk space to allocate to image
storage. We also stand to benefit by
dropping significant complexity from the
codebase.