limit raster tile rendering concurrency
Limit concurrency to the specified level
and render tiles in LIFO order. This makes
tiles more responsive.

Discard oldest items if the tile queue
gets too large since these are unlikely
to be useful. If the user scrolls quickly
we can end up with a large backlog of
wasteful work. The queue limit is somewhat
arbitrary, in the future we may want to
improve the algorithm by checking to see
if requested tiles are in the visible
region.