fix omitted layers with multiple sources
Fix an issue that caused incorrect sources to be
used occasionally when rendering a tile. The issue
is intermittent, depending on timing.

The problem was caused by deduplication keys that
did not account for the sources used in a job.
By having incorrect deduplications keys, a job
could complete with the results of a different
job for the same tile. The issue is fixed by
adding sources to the deduplication key.